Modularizing file to deal with Redis updates
